Data Validation: Using Model State and annotations(Route Constraints)

+++ api/Controllers/CommentController.cs
@@ -23,21 +23,27 @@ namespace api.Controllers{
         }
         [HttpGet]
         public async Task<IActionResult> GetAll(){
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var comments = await _commentRepository.GetAllAsync();
             var commmentsDto = comments.Select(c => c.ToCommentDto());
             return Ok(commmentsDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id){
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var commment = await _commentRepository.GetByIdAsync(id);
             if(commment == null){
                 return NotFound();
             }
             return Ok(commment.ToCommentDto());
         }
-        [HttpPost("{id}")]
+        [HttpPost("{id:int}")]
         public async Task<IActionResult> CreateComment([FromRoute] int id, [FromBody] CreateCommentDto commentDto){
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             if(!await _stockRepository.StockExists(id)){
                 return BadRequest("Stock does not exist");
             }
@@ -47,8 +53,10 @@ namespace api.Controllers{
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto){
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var commment = await _commentRepository.UpdateAsync(id, updateDto.ToCommentFromUpdate());
             if(commment == null){
                 return NotFound("Comment Not Found");
@@ -56,10 +64,13 @@ namespace api.Controllers{
             return Ok(commment.ToCommentDto());
         }
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+            
             var commment = await _commentRepository.DeleteAsync(id);
             if(commment == null) return NotFound("Comment Not Found");
+            return Ok(commment.ToCommentDto());
         }
     }
 }
